+++ src/homepage.js
@@ -1,49 +1,38 @@
 function removeobj(obj) {
-	$('.backgroundgreen').eq(obj).remove();
+	$('.gardeninfobox').eq(obj).remove();
 }
 
-function hidenofood() { // hides gardens on home page with no food
-	var objects = $('.backgroundgreen')
+
+function hidenofood() {
+	var objects = $('.gardeninfobox')
+	var wilting = "-48px 0px";
+	var food = "-160px 0px";
+	var frozen = "-96px 0px"
 	for (var i = objects.length - 1; i > -1; i--) {
-		var wilting = "-48px 0px";
-		var food = "-160px 0px";
-		var frozen = "-96px 0px"
-		var stat = $('.backgroundgreen').eq(i).children().eq(1).children().eq(2).children().eq(1).children().eq(0).css("background-position");
-		if (stat == wilting) {
-			continue;
-		}
-		if (stat == frozen) { // frozen
+		var iconlist = objects.eq(i).children().eq(0).children().eq(1).children().eq(1).children().eq(0).children();
+		if (iconlist.eq(0).css("background-position") == frozen) {
 			removeobj(i);
-			continue;
-		}
-		var totalobjs = $('.backgroundgreen').eq(i).children().eq(1).children().eq(2).children().eq(1).children().length;
-		if (totalobjs == 0) {
 			continue;
 		}
 		
-		if (totalobjs == 1) {
+		if (iconlist.length == 1) {
 			removeobj(i);
-		} else if (totalobjs == 2) {
-			var pos = $('.backgroundgreen').eq(i).children().eq(1).children().eq(2).children().eq(1).children().eq(1).css("background-position");
-			if (pos == food) { // food in 2nd array
+			continue;
+		} else if (iconlist.length == 2) {
+			if (iconlist.eq(1).css("background-position") == food) {
 				continue;
 			} else {
 				removeobj(i);
 				continue;
 			}
-			
-		} else if (totalobjs == 3) { 
-			var pos2 = $('.backgroundgreen').eq(i).children().eq(1).children().eq(2).children().eq(1).children().eq(2).css("background-position");
-			if (pos2 == food) { // food in 3rd array
+		} else if (iconlist.length == 3) {
+			if (iconlist.eq(2).css("background-position") == food) {
 				continue;
 			} else {
 				removeobj(i);
 				continue;
 			}
-		} else {
-			continue;
 		}
-		
 	}
 }
 
@@ -71,6 +60,13 @@ function openhomepageurls(type) { // handles "open all" button on wildlife page
 		window.open(links[i]);
 	}
 }
+function gardensearchreorder() {
+	var gardeninfoboxes = $('.gardeninfobox')
+	$('.gardeninfobox').eq(gardeninfoboxes.length-1).appendTo('#searchgardendiv')
+	$('#searchgardendiv').eq(0).insertAfter('#openallrandoms')
+}
+
+
 
 function addmainpagebuttons() {
 	$("div[style='color:#333;font-size:14px;font-weight:bold;width:220px;padding:2px;border-radius:5px']").eq(0).after('<input id="openallfriends" class="inputsubmit" style="width:194px;border:1px outset #0a2d0a;background-color:#007f00;padding-top: 8px;padding-bottom: 8px;" type=button value="Open All"/>')
@@ -80,18 +76,30 @@ function addmainpagebuttons() {
 	$('div[style="color:#333;font-size:14px;font-weight:bold;width:220px;padding:2px;border-radius:5px"]').eq(1).parent().attr("id","homepagefavorites");
 
 	$("div[style='color:#333;font-size:14px;font-weight:bold;width:220px;padding:2px;border-radius:5px']").eq(2).after('<input id="openallrandoms" class="inputsubmit" style="width:194px;border:1px outset #0a2d0a;background-color:#007f00;padding-top: 8px;padding-bottom: 8px;" type=button value="Open All"/>')
-	$('div[style="color:#333;font-size:14px;font-weight:bold;width:220px;padding:2px;border-radius:5px"]').eq(2).parent().attr("id","homepagerandoms");
+	//$('div[style="color:#333;font-size:14px;font-weight:bold;width:220px;padding:2px;border-radius:5px"]').eq(2).parent().attr("id","homepagerandoms");
 
 	document.getElementById ("openallfriends").addEventListener ("click", openhomepageurls.bind(this, "friends"));
+	if(document.getElementById ("openallfavorites")) {
 	document.getElementById ("openallfavorites").addEventListener ("click", openhomepageurls.bind(this, "favorites"));
 	document.getElementById ("openallrandoms").addEventListener ("click", openhomepageurls.bind(this, "randoms"));
+	}
 }
 
 function mainpage(settings) {
 	if (settings.hidenofood) {
 		hidenofood();
 	}
+	var path = window.location.pathname + window.location.search
+	if (path != "/fbfairy/index.php?mode=wilt" && path != "/fbfairy/index.php?mode=new") {
+		$('div[style="color:#333;font-size:14px;font-weight:bold;width:220px;padding:2px;border-radius:5px"]').eq(2).parent().attr("id","homepagerandoms");
+		$('div[style="color:#333;font-size:14px;font-weight:bold;width:220px;padding:2px;border-radius:5px"]').last().attr("id","searchgardendiv");
+	}
 	if (settings.homepageExtraBtns) {
 		addmainpagebuttons();
+	}
+	if (path != "/fbfairy/index.php?mode=wilt" && path != "/fbfairy/index.php?mode=new") {
+		if (settings.homepageReorderSearch) {
+			gardensearchreorder();
+		}
 	}
 }